* 'master' of https://github.com/xinzhiyun/xinpin:
  修改个人信息数据
  修改搜索样式
  新增：删除经销商时，有设备不能删除
  修改IC卡绑定功能
  新增管理员修改经销商密码
  修改个人信息编辑样式

+++ preject/Application/Admin/Controller/VendorsController.class.php
@@ -101,14 +101,18 @@ class VendorsController extends CommonController
         if ($userinfo[0]['leavel'] == 0 ) {
             $this->error('不能删除超级管理员！');
         }else{
-            // 查
-            $res = D('vendors')->delete($id);
-            if($res){
-                $this->success('删除成功',U('index'));
+            // 查询该经销商是否已经绑定设备，如绑定，则不允许删除
+            $binding = M('binding')->where("vid=".$id)->find();
+            if (!$binding) {
+                $res = D('vendors')->delete($id);
+                if($res){
+                    $this->success('删除成功',U('index'));
+                }else{
+                    $this->error('删除失败');
+                }
             }else{
-                $this->error('删除失败');
+                $this->error('有设备挂在该经销商名下，不能删除！');    
             }
-
         }
     }
 
@@ -280,6 +284,40 @@ class VendorsController extends CommonController
             }
 
         }else{
+            $this->display();
+        }
+    }
+
+
+    /**
+     * 修改经销商密码方法
+     * 
+     * @author 潘宏钢 <619328391@qq.com>
+     */
+    public function editpwd($id)
+    {
+        if (IS_POST) {
+            $new = md5(I('newpassword')); 
+            $re = md5(I('repassword'));
+            $id = I('id');
+
+            if ($new == $re) {
+                $user = M('Vendors');
+                
+                $res = $user->where('id='.$id)->setField('password',$new);
+                if ($res) {
+                    $this->success('修改经销商密码成功',U('index'));
+                }else{
+                    $this->error('修改密码失败！');
+                }
+               
+            }else{
+                $this->error('两次密码不一样，请重新输入！');
+            }
+
+        }else{
+            $info[] = D('vendors')->find($id);
+            $this->assign('info',$info);
             $this->display();
         }
     }
